+++ controller/server/view/network_list_page.ml
@@ -3,7 +3,8 @@ open Tyxml.Html
 open Sexplib.Std
 open Protocol_conv_jsonm
 
-let service_item ({ id; name; strength; ipv4 } as service) =
+let service_item interface_annotations ({ id; name; strength; ipv4 } as service)
+    =
   let icon =
     match strength with
     | Some s ->
@@ -21,7 +22,19 @@ let service_item ({ id; name; strength; ipv4 } as service) =
   li
     [ a
         ~a:[ a_class classes; a_href ("/network/" ^ id) ]
-        [ div [ txt name ]
+        [ div
+            [ txt name
+            ; span
+                ~a:[ a_class [ "d-NetworkList__Labels" ] ]
+                (List.assoc_opt service.ethernet.interface interface_annotations
+                |> Option.value ~default:[]
+                |> List.map (fun label_text ->
+                       span
+                         ~a:[ a_class [ "d-NetworkList__Label" ] ]
+                         [ txt label_text ]
+                   )
+                )
+            ]
         ; ( match ipv4 with
           | Some ipv4_addr ->
               div
@@ -39,10 +52,11 @@ type params =
   { proxy : string option
   ; services : Connman.Service.t list
   ; interfaces : Network.Interface.t list
+  ; interface_annotations : (string * string list) list
   }
 [@@deriving protocol ~driver:(module Jsonm)]
 
-let html { proxy; services; interfaces } =
+let html { proxy; services; interfaces; interface_annotations } =
   let connected_services, available_services =
     List.partition Connman.Service.is_connected services
   in
@@ -64,7 +78,10 @@ let html { proxy; services; interfaces } =
              section
                [ ul
                    ~a:[ a_class [ "d-NetworkList" ]; a_role [ "list" ] ]
-                   (List.map service_item connected_services)
+                   (List.map
+                      (service_item interface_annotations)
+                      connected_services
+                   )
                ]
          )
        ; Definition.list
@@ -96,7 +113,7 @@ let html { proxy; services; interfaces } =
                else
                  ul
                    ~a:[ a_class [ "d-NetworkList" ]; a_role [ "list" ] ]
-                   (List.map service_item available_services)
+                   (List.map (service_item []) available_services)
              )
            ]
        ; section